Move PayoutDetails to /payments-tab
